build: package layout and test conversion to TS
