+++ lib/index.d.ts
@@ -1,7 +1,7 @@
 export function is_object(obj: unknown): boolean;
 export function is_object_literal(obj: unknown): boolean;
-export function compare(el1: Object, el2: Object): boolean;
-export function clone<T extends object>(target: T): T;
+export function compare(el1: unknown, el2: unknown): boolean;
+export function clone<T>(target: T): T;
 
 type Merge<T extends object[]> = T extends [
   infer Head extends object,
@@ -17,12 +17,32 @@ type Merge<T extends object[]> = T extends [
       >
   : never;
 
+// type MergeObjects<T extends object, U extends object> = {
+//   [K in keyof T | keyof U]: K extends keyof T & keyof U
+//     ? T[K] extends object
+//       // Could simplify the code but type negation doesn't seem to work, see
+//       // https://github.com/microsoft/TypeScript/issues/4196
+//       // https://github.com/microsoft/TypeScript/pull/29317
+//       // https://github.com/microsoft/TypeScript/issues/4196
+//       ? U[K] extends object & !Buffer
+//         ? MergeObjects<T[K], U[K]>
+//         : T[K]
+//       : U[K]
+//     : K extends keyof T
+//     ? T[K]
+//     : K extends keyof U
+//     ? U[K]
+//     : never;
+// };
+
 type MergeObjects<T extends object, U extends object> = {
   [K in keyof T | keyof U]: K extends keyof T & keyof U
     ? T[K] extends object
-      ? U[K] extends object
-        ? MergeObjects<T[K], U[K]>
-        : T[K]
+      ? U[K] extends Buffer
+        ? T[K]
+        : U[K] extends object
+          ? MergeObjects<T[K], U[K]>
+          : T[K]
       : U[K]
     : K extends keyof T
     ? T[K]
@@ -48,5 +68,5 @@ type SnakeCaseKeys<T> = T extends Array<infer U>
 
 export function snake_case<T extends Record<string, any>>(
   source: T,
-  convert?: boolean
+  convert?: boolean | number
 ): SnakeCaseKeys<T>;
