+++ package.json
@@ -11,6 +11,8 @@
     "@babel/preset-env": "^7.15.6",
     "@commitlint/cli": "^13.1.0",
     "@commitlint/config-conventional": "^13.1.0",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.15.3",
     "coffeescript": "^2.6.0",
     "husky": "^7.0.2",
     "mocha": "^9.1.1",
@@ -18,7 +20,9 @@
     "rollup": "^2.56.3",
     "rollup-plugin-babel": "^4.4.0",
     "should": "^13.2.3",
-    "standard-version": "^9.3.1"
+    "standard-version": "^9.3.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   },
   "engines": {
     "node": ">= 8.0.0"
@@ -26,6 +30,18 @@
   "homepage": "https://github.com/adaltas/node-mixme",
   "bugs": {
     "url": "https://github.com/adaltas/node-mixme/issues"
+  },
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./lib/index.d.ts",
+        "default": "./lib/index.js"
+      },
+      "require": {
+        "types": "./dist/cjs/index.d.cts",
+        "default": "./dist/cjs/index.cjs"
+      }
+    }
   },
   "keywords": [
     "clone",
@@ -40,17 +56,17 @@
     "/dist"
   ],
   "license": "MIT",
-  "main": "dist/mixme.cjs.js",
-  "module": "dist/mixme.esm.js",
+  "main": "dist/cjs/mixme.js",
   "mocha": {
-    "throw-deprecation": true,
+    "inline-diffs": true,
+    "loader": "ts-node/esm",
+    "recursive": true,
+    "reporter": "spec",
     "require": [
       "should"
     ],
-    "inline-diffs": true,
-    "timeout": 40000,
-    "reporter": "spec",
-    "recursive": true
+    "throw-deprecation": true,
+    "timeout": 40000
   },
   "repository": {
     "type": "git",
@@ -60,13 +76,14 @@
     "postinstall": "husky install",
     "prepublishOnly": "pinst --disable",
     "postpublish": "pinst --enable",
-    "build": "rollup -c",
-    "test": "mocha 'test/**/*.js'",
+    "build": "rollup -c && cp lib/index.d.ts dist/cjs/index.d.cts",
+    "test": "mocha 'test/**/*.{js,ts}'",
     "release": "standard-version",
     "release:minor": "standard-version --release-as minor",
     "release:patch": "standard-version --release-as patch",
     "release:major": "standard-version --release-as major"
   },
   "type": "module",
-  "types": "index.d.ts"
+  "types": "lib/index.d.ts",
+  "dependencies": {}
 }
